fusion of the welcome.py and logging.py
